requirements.txt
uwsgi==2.0.18
psycopg2==2.8.2
django==2.2
djangorestframework==3.9.2
django-filter==2.1.0